new try

+++ hashmap.c
@@ -126,7 +126,7 @@ Pair * firstMap(HashMap * map) {
 }
 
 Pair * nextMap(HashMap * map) {
-  for(long i = 0; i < map->capacity;i++){
+  for(int i = 0; i < map->capacity;i++){
     struct Pair* currentpair = map->buckets[i];
     if(currentpair != NULL && currentpair->key != NULL){
       map->current = i;
